requirements.txt
Flask
python-dotenv
Flask-WTF
email-validator
Flask-Login
Flask-PyMongo
Authlib
Werkzeug